requirements.txt
robotframework>=7
robotframework-robocop>=6.0.0
chromedriver-py==134.*
QWeb>=3.5.0
robotframework-libtoc==1.4.2